+++ src/components/Card/Card.tsx
@@ -1,18 +1,14 @@
 import { useState } from "react";
-import  { type CardProps, View } from "./CardType";
+import { type CardProps, View } from "./CardType";
 import rupee from "../../assets/rupee.png";
 import dayjs from "dayjs";
+import ScrollToTopButton from "../ScrollToTopButton/ScrollToTopButton";
 
 export default function Card({ products }: CardProps) {
   const [toggle, setToggle] = useState<View>(View.Grid);
-  function handleScollToTop(){
-    window.scrollTo({
-      top:0,
-      behavior:"smooth"
-    })
-  }
+
   return (
-    <div className="flex flex-col items-center p-2 w-full box-border">
+    <div className="ml-64 mt-16 flex flex-col items-center p-4 w-full box-border">
       <div className="flex gap-2 m-3">
         <button
           className={`mt-1 right-1 z-10 border px-4 py-2 rounded  ${
@@ -58,7 +54,7 @@ export default function Card({ products }: CardProps) {
           </div>
         ))}
       </div>
-      <button onClick={handleScollToTop} className="fixed p-2 bg-blue-400 rounded-xl  bottom-4 right-4">SKip To Top</button>
+      <ScrollToTopButton />
     </div>
   );
 }
