+++ src/utility/debounce.ts
@@ -1,0 +1,15 @@
+export default function debounce<T extends (...args: any[]) => void>(
+    fn: T,
+    delay: number
+  ) {
+    let timer: ReturnType<typeof setTimeout> | null = null;
+    return (...args: Parameters<T>) => {
+      if (timer) {
+        clearTimeout(timer);
+      }
+      timer = setTimeout(() => {
+        fn(...args);
+      }, delay);
+    };
+  }
+
